feat: add api client hook
