+++ frontend/src/services/useApiClient.js
@@ -1,0 +1,7 @@
+import { useMemo } from 'react';
+import { APIClient } from './api-client.js';
+
+// Returns an API client instance scoped to the current auth token
+export default function useApiClient() {
+  return useMemo(() => new APIClient(), []);
+}
